Fix downsampling level selection and slide dimensions for improved cross level pixel correspondance.
Fixs: 1) Lines 190 - 214 Improved target level selection

Openslide reports the dimensional downsampling of layers as being a float level0/desired level width. Do to integer arithmetic these may not be multiples of each other.

As an example: 10x acquisitions, report being downsampled ~4.0001x from 40x due to this.  As a result, calling openslide_get_best_level_for_downsample and requesting a 4x downsample. Openslide will return the 40x image and not the 10x  actual 10x acquisition.  To fix this the code now just compares the dimensions of the downsample with the target downsample dimensions and finds the layer which has the dimensions which best match desired level dimensions

Fix 2) Line 229 Taking floor downsampled target level

const double multiplicator =    floor(openslide_get_level_downsample(osr, levelToGet));

multiplicator = downsampling to go from Level0 -> the optimal layer selected for pulling images from.

Taking the floor of the value returned by openslide corrects for non integer downsampling reported in in aspire imaging.

Fix 3: Lines 265 - 334

For down sampled images which are not direct multiples of each other. The hight and width of the images changes.  Due to pixels not captured in the downsampling. This changes the pixels spacing of the downsampled images.  These lines calculate the functional change in the dimensions of the downsampled images to report the true dimensions of the generated downsample images.